Figure out the coordinate systems of accel/gyro/mag and their transformation.
According to datasheet, coordinates on chip MPU9150 are:
accel/gyro: right-front-up; mag: front-right-down.
Due to the layout of MPU9150 on the board, there is a 135 degree difference, so a coordinate transformation is needed.
It is realized by transformation() and transformation_mag() in util.c. These 2 functions are called in ERU_Event_Handler() in MPU9150.c.
After the transformation, we get identical board coordinate system for accel/gyro/mag: front-left-up.

CAUTIOUS: acceleration is calculated according to external force!
when put the board static and upward, external force is upward. z-axis accel value is G, so z-axis positive direction is upward.
SO when put static, if the positive direction of any axis is consistent with gravity, accel value should be -G!!!