Succeeded to get heading angle. Update mag bias.
1. Get heading angle using only mag data with GetHeading().
NOTICE that sin() and cos() use radians rather than degrees!!!!!!!
Remove GetYaw().

2. Add codes in ERU_Event_Handler(void) in MPU9150.c to calculate mag bias, and update bias in transformation_mag() in util.c.
The theory is: rotate the board and record the max and min mag value of each axis. Bias is the average of max and min. Print out bias value using serial port, then finally update the bias[3] in util.c manually.
It is possible to encapsulate mag calibration into a function, but not necessary for now. So far I don't know how to save calibration value to Flash or ROM.